Add G3 PATCH_NAME changes dettection. CommonCause.PARAM_CHANGED -> CommonCause.PARAM_VALUE. CommonCause.EFFECT_CHANGED -> CommonCause.EFFECT_VALUE. Simplified ZoomGSeriesActiveEffectDecoder

+++ MultistompEditor/multistomp/br/com/srmourasilva/multistomp/zoom/gseries/decoder/effect/ZoomGSeriesPatchEffectsStatusDecoder.java
@@ -34,7 +34,7 @@ public class ZoomGSeriesPatchEffectsStatusDecoder implements MessageDecoder {
 	}
 
 	private boolean hasActived(MidiMessage message, int position) {
-		return SysexVerbal.For(message).asBoolean(position, (byte) 0x0);
+		return SysexVerbal.For(message).asBoolean(position, 0x0);
 	}
 
     private Message generateMessageFor(int effect, boolean actived) {
